Fejl med audio under review
Review Environment

Submission ID: acd4da07-94ff-456d-b25f-c7aa8d7da746
Review date: May 22, 2025
Version reviewed: 1.0

Guideline 2.5.4 - Performance - Software Requirements

The app declares support for audio in the UIBackgroundModes key in your Info.plist, but we are unable to play any audible content when the app is running in the background.

Background audio is intended for use by apps that provide audible content to the user while in the background, such as music player, music creation, or streaming audio apps.

Next Steps

If the app has a feature that requires persistent audio, reply to this message and let us know how to locate this feature. If the app does not have a feature that requires persistent audio, it would be appropriate to remove the "audio" setting from the UIBackgroundModes key.